da xog prductbatch ( nhưng chưa fix kèm adjust)

+++ AgriculturalWarehouseManagement/src/main/java/com/example/AgriculturalWarehouseManagement/Backend/repositorys/WarehouseRepository.java
@@ -3,5 +3,6 @@ package com.example.AgriculturalWarehouseManagement.Backend.repositorys;
 import com.example.AgriculturalWarehouseManagement.Backend.models.Warehouse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
+
+public interface WarehouseRepository extends JpaRepository<Warehouse, Integer> {
 }
